requirements.txt
altgraph==0.17.3
asttokens==2.2.1
backcall==0.2.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
emoji==1.7.0
entrypoints==0.4
executing==1.2.0
firebase-admin==6.0.1
fonttools==4.38.0
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
msgpack==1.0.4
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.5
packaging==21.3
parso==0.8.3
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.14
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pywin32==305
pywin32-ctypes==0.2.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
rsa==4.9
six==1.16.0
SpeechRecognition==3.8.1
stack-data==0.6.2
text2emotion==0.0.5
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
